video-service/pyproject.toml
[project]
name = "video-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Максим",email = "maksim_goncharov_2006@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pydantic[email] (>=2.11.4,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)"
]

[tool.poetry]
packages = [{include = "video_service", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
